+++ Clockify.Net/Models/Estimates/EstimateRequest.cs
@@ -1,8 +1,11 @@
-﻿namespace Clockify.Net.Models.Estimates
+﻿using Newtonsoft.Json;
+
+namespace Clockify.Net.Models.Estimates
 {
     public class EstimateRequest
     {
         public int? Estimate { get; set; }
+
         public EstimateType Type { get; set; }
     }
 }