+++ Clockify.Tests/Tests/TagTests.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Clockify.Net;
 using Clockify.Net.Models.Tags;
 using Clockify.Net.Models.Workspaces;
-using Clockify.Tests.Fixtures;
+using Clockify.Tests.Helpers;
 using FluentAssertions;
 using NUnit.Framework;
 
@@ -22,21 +22,21 @@ namespace Clockify.Tests.Tests
         [OneTimeSetUp]
         public async Task Setup()
         {
-            var workspaceResponse = await _client.CreateWorkspaceAsync(new WorkspaceRequest {Name = "TagsWorkspace"});
-            workspaceResponse.IsSuccessful.Should().BeTrue();
-            _workspaceId = workspaceResponse.Data.Id;
+	        _workspaceId = await SetupHelper.CreateOrFindWorkspaceAsync(_client, "Clockify.NetTestWorkspace");
         }
 
-        [OneTimeTearDown]
-        public async Task Cleanup()
-        {
-	        var currentUser = await _client.GetCurrentUserAsync();
-	        var changeResponse =
-		        await _client.SetActiveWorkspaceFor(currentUser.Data.Id, DefaultWorkspaceFixture.DefaultWorkspaceId);
-	        changeResponse.IsSuccessful.Should().BeTrue();
-            var workspaceResponse = await _client.DeleteWorkspaceAsync(_workspaceId);
-            workspaceResponse.IsSuccessful.Should().BeTrue();
-        }
+        // TODO Uncomment when Clockify add deleting workspaces again
+
+        //[OneTimeTearDown]
+        //public async Task Cleanup()
+        //{
+	       // var currentUser = await _client.GetCurrentUserAsync();
+	       // var changeResponse =
+		      //  await _client.SetActiveWorkspaceFor(currentUser.Data.Id, DefaultWorkspaceFixture.DefaultWorkspaceId);
+	       // changeResponse.IsSuccessful.Should().BeTrue();
+        //    var workspaceResponse = await _client.DeleteWorkspaceAsync(_workspaceId);
+        //    workspaceResponse.IsSuccessful.Should().BeTrue();
+        //}
 
         [Test]
         public async Task FindAllTagsOnWorkspaceAsync_ShouldReturnTagsList()
@@ -50,7 +50,7 @@ namespace Clockify.Tests.Tests
         {
             var tagRequest = new TagRequest
             {
-                Name = "Test tag",
+                Name = "Test tag " + Guid.NewGuid(),
             };
             var createResult = await _client.CreateTagAsync(_workspaceId, tagRequest);
             createResult.IsSuccessful.Should().BeTrue();