+++ Clockify.Tests/Tests/UserTests.cs
@@ -2,7 +2,7 @@ using System.Net;
 using System.Threading.Tasks;
 using Clockify.Net;
 using Clockify.Net.Models.Workspaces;
-using Clockify.Tests.Fixtures;
+using Clockify.Tests.Helpers;
 using FluentAssertions;
 using NUnit.Framework;
 
@@ -21,22 +21,22 @@ namespace Clockify.Tests.Tests
         [OneTimeSetUp]
         public async Task Setup()
         {
-            var workspaceResponse =
-                await _client.CreateWorkspaceAsync(new WorkspaceRequest {Name = "UserTest Workspace"});
-            workspaceResponse.IsSuccessful.Should().BeTrue();
-            _workspaceId = workspaceResponse.Data.Id;
+            _workspaceId = await SetupHelper.CreateOrFindWorkspaceAsync(_client, "Clockify.NetTestWorkspace");
+
         }
 
-        [OneTimeTearDown]
-        public async Task Cleanup()
-        {
-	        var currentUser = await _client.GetCurrentUserAsync();
-	        var changeResponse =
-		        await _client.SetActiveWorkspaceFor(currentUser.Data.Id, DefaultWorkspaceFixture.DefaultWorkspaceId);
-	        changeResponse.IsSuccessful.Should().BeTrue();
-            var workspaceResponse = await _client.DeleteWorkspaceAsync(_workspaceId);
-            workspaceResponse.IsSuccessful.Should().BeTrue();
-        }
+        // TODO Uncomment when Clockify add deleting workspaces again
+
+        //[OneTimeTearDown]
+        //public async Task Cleanup()
+        //{
+	       // var currentUser = await _client.GetCurrentUserAsync();
+	       // var changeResponse =
+		      //  await _client.SetActiveWorkspaceFor(currentUser.Data.Id, DefaultWorkspaceFixture.DefaultWorkspaceId);
+	       // changeResponse.IsSuccessful.Should().BeTrue();
+        //    var workspaceResponse = await _client.DeleteWorkspaceAsync(_workspaceId);
+        //    workspaceResponse.IsSuccessful.Should().BeTrue();
+        //}
 
         [Test]
         public async Task GetCurrentUser_ShouldReturnCurrentUser()