+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.0]
+### Added
+ - Support for deleting and updating tasks. (PR #16)
+ - `Note` field to ProjectRequest and ProjectDtoImpl (PR #15)
+ - `FindAllHydratedTimeEntriesForUserAsync` method to client (PR #12)
+### Changed
+- JsonSerializer to Json.Net
+### Fixed
+- `EstimateRequest` should now work properly. 
+
 ## [1.1.4]
 ### Fixed
 - `Estimate` deserialization (PR #10)